refactor: 상수값 외부 파일로 분리
