+++ config-overrides.js
@@ -9,6 +9,7 @@ module.exports = override(
     '@routes': path.resolve(__dirname, 'src/routes'),
     '@api': path.resolve(__dirname, 'src/api'),
     '@components': path.resolve(__dirname, 'src/components'),
+    '@constants': path.resolve(__dirname, 'src/constants'),
     '@assets': path.resolve(__dirname, 'src/assets'),
     '@pages': path.resolve(__dirname, 'src/pages'),
     '@styles': path.resolve(__dirname, 'src/styles'),
