+++ src/constants/transportation.ts
@@ -1,15 +1,15 @@
 const BUS = '버스';
 const SUBWAY = '지하철';
-const TEXI = '택시';
+const TAXI = '택시';
 const TRAIN = 'KTX/기차';
 const WALKING = '도보';
 const ELECTRIC_SCOOTER = '전동킥보드';
 const CAR = '자가용';
 
-export const transportaions = [
+export const transportations = [
   BUS,
   SUBWAY,
-  TEXI,
+  TAXI,
   TRAIN,
   WALKING,
   ELECTRIC_SCOOTER,
