+++ tsconfig.paths.json
@@ -6,6 +6,7 @@
       "@routes/*": ["src/routes/*"],
       "@api/*": ["src/api/*"],
       "@components/*": ["src/components/*"],
+      "@constants/*": ["src/constants/*"],
       "@assets/*": ["src/assets/*"],
       "@pages/*": ["src/pages/*"],
       "@styles/*": ["src/styles/*"],
